fix buttons style buttons components
