+++ src/screens/audio/components/Table.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import ButtonSound from "../../../ui-library/ButtonSound";
+import ButtonSound from "../../../ui-library/buttons/ButtonSound";
 import useSound from "use-sound";
 
 const Table = (props: { k?: any; v?: any }) => {
